Refactor for better file structure
