+++ src/main/java/com/demo/filmdb/config/SpringDocConfig.java
@@ -1,4 +1,4 @@
-package com.demo.filmdb.utils;
+package com.demo.filmdb.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
